OpenAIとのAPI通信部分をサービスクラスに分離する
